game_server/requirements.txt
# ===================================================================
# Основной фреймворк (Web Framework)
# ===================================================================
fastapi
uvicorn
gunicorn


# ===================================================================
# Фоновые задачи (Background Tasks) - ЗАМЕНА НА ARQ
# ===================================================================
arq
redis

# ===================================================================
# Работа с данными (Data & Databases) - БЕЗ ИЗМЕНЕНИЙ
# ===================================================================
# ORM и миграции
SQLAlchemy
alembic
Mako

# Драйверы для PostgreSQL (синхронный и асинхронный)
psycopg2-binary
asyncpg

# Драйвер для Redis (синхронный и асинхронный)
aioredis  # Сохраняем для обратной совместимости

# Работа с YAML файлами (для сидов)
PyYAML

# ===================================================================
# Безопасность (Security) - БЕЗ ИЗМЕНЕНИЙ
# ===================================================================
# Хэширование паролей
passlib[bcrypt]
bcrypt

# JWT-токены
PyJWT

# ===================================================================
# Интеграции и внешние API (Integrations & APIs)
# ===================================================================
# HTTP-клиенты (синхронный и асинхронный)
requests
aiohttp
websockets

# API для Discord
disnake  # Изменено с discord.py на disnake

# Сериализация
msgpack # Добавлено

# ===================================================================
# Мониторинг и Наблюдаемость (Monitoring & Observability) - БЕЗ ИЗМЕНЕНИЙ
# ===================================================================
prometheus-client
prometheus-fastapi-instrumentator

# ===================================================================
# Инструменты и утилиты (Tooling & Utilities) - БЕЗ ИЗМЕНЕНИЙ
# ===================================================================
# Загрузка .env файлов
python-dotenv

# Логирование
loguru
colorlog

# Системная информация
psutil

# Парсинг SQL
sqlparse

# Библиотека для умных повторных попыток
tenacity

# ===================================================================
# Тестирование и типизация (Testing & Typing) - БЕЗ ИЗМЕНЕНИЙ
# ===================================================================
pytest
rich
typing_extensions

# uvloop - уже был, оставлен для явности и производительности asyncio
uvloop
aio-pika  # Сохраняем, если используется в других частях приложения

Pydantic
NumPy

numba

aio-pika
pymongo
motor
inject